+++ src/path_planner/compound.rs
@@ -10,9 +10,9 @@ use vexide::float::Float as _;
 /// that are connected together. The paths must be connected at their endpoints.
 pub struct CompoundPath {
     paths: Vec<Box<dyn Path>>,
-    lengths: Vec<f32>, // Cumulative lengths of the paths
-    total_length: f32,
-    path_t: f32,
+    lengths: Vec<f64>, // Cumulative lengths of the paths
+    total_length: f64,
+    path_t: f64,
 }
 
 impl CompoundPath {
@@ -41,7 +41,7 @@ impl CompoundPath {
         }
 
         Self {
-            path_t: 1.0 / paths.len() as f32,
+            path_t: 1.0 / paths.len() as f64,
             paths,
             lengths,
             total_length,
@@ -50,13 +50,13 @@ impl CompoundPath {
 }
 
 impl Path for CompoundPath {
-    fn length_until(&self, t: f32) -> f32 {
+    fn length_until(&self, t: f64) -> f64 {
         // If t is 1.0, we are at the end of the last path
         if t == 1.0 {
             return self.total_length;
         }
         // Find which path t is in
-        let path = (self.paths.len() as f32 * t).floor() as isize;
+        let path = (self.paths.len() as f64 * t).floor() as isize;
         // If path is less than 0, use the first path. If it's more than the
         // number of paths, use the last path.
         let path = if path < 0 {
@@ -67,7 +67,7 @@ impl Path for CompoundPath {
             path as usize
         };
         // Find t along that path
-        let local_t = (t - path as f32 * self.path_t) / self.path_t;
+        let local_t = (t - path as f64 * self.path_t) / self.path_t;
         self.paths[path].length_until(local_t)
             + if path > 0 {
                 self.lengths[path - 1]
@@ -76,9 +76,9 @@ impl Path for CompoundPath {
             }
     }
 
-    fn evaluate(&self, t: f32) -> Pose {
+    fn evaluate(&self, t: f64) -> Pose {
         // Find which path t is in
-        let path = (self.paths.len() as f32 * t).floor() as isize;
+        let path = (self.paths.len() as f64 * t).floor() as isize;
         // If path is less than 0, use the first path. If it's more than the
         // number of paths, use the last path.
         let path = if path < 0 {
@@ -89,7 +89,7 @@ impl Path for CompoundPath {
             path as usize
         };
         // Find t along that path
-        let local_t = (t - path as f32 * self.path_t) / self.path_t;
+        let local_t = (t - path as f64 * self.path_t) / self.path_t;
         // Evaluate the path at the given t
         self.paths[path].evaluate(local_t)
     }
